+++ src/spyglass/position/v1/dlc_utils_makevid.py
@@ -422,8 +422,17 @@ class VideoMaker:
         if self.debug:
             print(f"\r{msg}", end=end)
 
-    def plot_frames(self, start_frame, end_frame, progress_bar=None):
+    def plot_frames(
+        self, start_frame, end_frame, progress_bar=None, process_pool=True
+    ):
         logger.debug(f"Plotting   frames: {start_frame} - {end_frame}")
+
+        if not process_pool:  # Single-threaded processing for debugging
+            for frame_ind in range(start_frame, end_frame):
+                self._generate_single_frame(frame_ind)
+                progress_bar.update()
+            return
+
         with ProcessPoolExecutor(max_workers=self.max_workers) as executor:
             jobs = {}  # dict of jobs