+++ src/spyglass/spikesorting/v0/spikesorting_curation.py
@@ -608,7 +608,9 @@ class QualityMetrics(SpyglassMixin, dj.Computed):
 
         for unit_id in waveform_extractor.sorting.get_unit_ids():
             # checks to avoid bug in spikeinterface 0.98.2
-            if num_spikes[unit_id] < min_spikes:
+            if num_spikes[unit_id] < min_spikes and (
+                is_nn_iso or is_nn_overlap
+            ):
                 if is_nn_iso:
                     metric[str(unit_id)] = (np.nan, np.nan)
                 elif is_nn_overlap: