+++ src/spyglass/utils/dj_mixin.py
@@ -158,9 +158,11 @@ class SpyglassMixin(ExportMixin):
             ):  # check if all required fields are in key
                 if not len(query := cls() & key) == 1:  # check if key is unique
                     raise KeyError(
-                        f"Key is neither fully specified nor a unique entry in"
-                        + f"table.\n\tTable: {cls.full_table_name}\n\tKey: {key}"
-                        + f"Required fields: {required_fields}\n\tResult: {query}"
+                        "Key is neither fully specified nor a unique entry in"
+                        + "table.\n\t"
+                        + f"Table: {cls.full_table_name}\n\tKey: {key}"
+                        + f"Required fields: {required_fields}\n\t"
+                        + f"Result: {query}"
                     )
                 key = query.fetch1("KEY")
 
@@ -174,10 +176,10 @@ class SpyglassMixin(ExportMixin):
 
         Used to determine fetch_nwb behavior. Also used in Merge.fetch_nwb.
         Implemented as a cached_property to avoid circular imports."""
-        from spyglass.common.common_nwbfile import (  # noqa F401
+        from spyglass.common.common_nwbfile import (
             AnalysisNwbfile,
             Nwbfile,
-        )
+        )  # noqa F401
 
         table_dict = {
             AnalysisNwbfile: "analysis_file_abs_path",
@@ -415,7 +417,7 @@ class SpyglassMixin(ExportMixin):
         if None in experimenters:
             raise PermissionError(
                 "Please ensure all Sessions have an experimenter in "
-                + f"SessionExperimenter:\n{sess_summary}"
+                + f"Session.Experimenter:\n{sess_summary}"
             )
 
         user_name = LabMember().get_djuser_name(dj_user)
@@ -848,8 +850,8 @@ class SpyglassMixin(ExportMixin):
                 "Connection ID",  # t.PROCESSLIST_ID -- User connection ID
                 "User",  # t.PROCESSLIST_USER -- User
                 "Host",  # t.PROCESSLIST_HOST -- User machine
-                "Process Database",  # t.PROCESSLIST_DB -- Thread database
                 "Time (s)",  # t.PROCESSLIST_TIME -- Time in seconds
+                "Process Database",  # t.PROCESSLIST_DB -- Thread database
                 "Process",  # t.PROCESSLIST_COMMAND -- Likely Query
                 "State",  # t.PROCESSLIST_STATE
                 "Query",  # t.PROCESSLIST_INFO -- Actual query